[Refactor] #244 JobPostingService Query, Command로 분리
